refactor(types): eliminate reservation types Middle Man files (X-Tracker #497)
- Removed completely redundant re-export files \src/components/reservation/types.ts\ and \src/services/reservations/types.ts\ to eliminate Middle Man code smell.
- Refactored all reservation components and hooks under \src/components/reservation/*\ and \src/hooks/user/reservation/*\ to import the \Reservation\ type directly from \@/types/reservation\.
- Restored service-level re-exports on actual service files (\countries.ts\, \	agCatalog.ts\, \mapMentor.ts\) for backwards compatibility, keeping them clean and highly performant.
- Verified all 800 tests pass and the production build completes flawlessly.
