--- src/components/reservation/types.ts
@@ -1,1 +1,0 @@
-export * from '@/types/reservation';
