--- src/services/reservations/types.ts
@@ -1,7 +1,0 @@
-import type {
-  MessageRole,
-  Reservation,
-  ReservationMessage,
-} from '@/types/reservation';
-
-export type { MessageRole, Reservation, ReservationMessage };
